Challenge 1 complete.

+++ schema.sql
@@ -1,14 +1,3 @@
--- Create a MySQL Database called bamazon.
--- Then create a Table inside of that database called products.
--- The products table should have each of the following columns:
-
---     item_id (unique id for each product)
---     product_name (Name of product)
---     department_name
---     price (cost to customer)
---     stock_quantity (how much of the product is available in stores)
-
--- Populate this database with around 10 different products. (i.e. Insert "mock" data rows into this database and table).
 DROP DATABASE IF EXISTS whamazon_db;
 CREATE DATABASE whamazon_db;
 
@@ -18,7 +7,7 @@ CREATE TABLE products(
     item_id INT AUTO_INCREMENT NOT NULL,
     product_name VARCHAR(50) NOT NULL,
     department_name VARCHAR(50) NOT NULL,
-    price FLOAT(4,2) DEFAULT 0,
+    price FLOAT(6,2) DEFAULT 0,
     stock_quantity INT(4) DEFAULT 0,
     PRIMARY KEY(item_id)
 );
